+++ internal/web/webgui.go
@@ -48,6 +48,7 @@ func Gui(dirPath, nodePath string) {
 
 	routerConf.GET("/", configHandler)              // config.go
 	routerConf.GET("/logout", logoutHandler)        // config.go
+	routerConf.GET("/config/notify", notifyHandler) // config.go
 	routerConf.POST("/", configHandler)             // config.go
 	routerConf.POST("/config/", saveConfigHandler)  // config.go
 	routerConf.POST("/config/auth", saveConfigAuth) // config.go